requirements.txt
# pip install torch torchvision torchaudio --index-url https://download.pytorch.org/whl/cu118
# pip install paddlepaddle-gpu --index-url https://mirror.baidu.com/pypi/simple


# Core Framework
langchain==0.2.15
langchain-community==0.2.15
langchain-core==0.2.37
langgraph==0.1.17
langchain-ollama==0.1.1
langchain-huggingface==0.0.3

# Embeddings & Vector Store
sentence-transformers==3.0.1
chromadb==0.5.23
faiss-cpu==1.8.0

# Image & Audio Processing
paddlepaddle
paddleocr==2.8.1
opencv-python==4.9.0.80
opencv-contrib-python==4.9.0.80
pillow==10.2.0
openai-whisper==20240930

# UI Framework
streamlit==1.40.1
streamlit-option-menu==0.4.0

# LLM & NLP
ollama==0.3.1
requests==2.32.3
python-dotenv==1.0.1

# Data & Database
sqlalchemy==2.0.35
pydantic==2.9.2
typing-extensions==4.12.2

# Utilities
numpy==1.26.4
pandas==2.2.3
python-dateutil==2.9.0

# Development & Testing
pytest==8.3.2
black==24.10.0
